Sidebar ahora se comporta como un layout para incluir interfaces sin la misma
