+++ src/View/Index.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Index () {
+  return (
+    <div>Index</div>
+  )
+}
+
+export default Index
