+++ src/View/Park.jsx
@@ -1,12 +1,16 @@
 import Slot from '../Component/ParkingSlot/Slot'
+import SideBar from '../Component/SideBar/sideBar'
+
 function Park () {
   return (
-    <div class='content-parking'>
-      <div class='content-btn'>
-        <div><button class='btn-distributio btn-parking' id='asignarParqueaderosButton'>Asignar Parqueaderos</button><button class='btn-parking' id='Save'>Guardar</button></div>
+    <SideBar>
+      <div class='content-parking'>
+        <div class='content-btn'>
+          <div><button class='btn-distributio btn-parking' id='asignarParqueaderosButton'>Asignar Parqueaderos</button><button class='btn-parking' id='Save'>Guardar</button></div>
+        </div>
+        <Slot numberOfSlots={32} />
       </div>
-      <Slot numberOfSlots={32} />
-    </div>
+    </SideBar>
   )
 }
